Fix warning about int to float wheelbase conversion

+++ cpp/core/configfile.cpp
@@ -10,7 +10,7 @@
 float config::V= 3.0; // m/s
 float config::MAXG= 30*pi/180; // radians, maximum steering angle (-MAXG < g < MAXG)
 float config::RATEG= 20*pi/180; // rad/s, maximum rate of change in steer angle
-int config::WHEELBASE= 4; // metres, vehicle wheel-base
+float config::WHEELBASE= 4.; // metres, vehicle wheel-base
 float config::DT_CONTROLS= 0.025; // seconds, time interval between control signals
 
 // control noises
